+++ src/Docs/View.fs
@@ -98,7 +98,8 @@ let private leftSide (p: Page) =
                         Daisy.menuTitle [ Html.span "Docs" ]
                         mi "Install" Install
                         mi "Use" Use
-                        mi "Slider" Slider
+                        mi "Examples" Examples
+                        mi "Advanced" Advanced
                         ]
                 ]
             ]
@@ -131,7 +132,8 @@ let AppView (state: State) (dispatch: Msg -> unit) =
         match state.Page with
         | Install -> "Installation", "/docs/install", Pages.Install.InstallView()
         | Use -> "How to use", "/docs/use", Pages.Use.UseView()
-        | Slider -> "Slider", "/docs/slider", Pages.Slider.SliderView()
+        | Examples -> "Examples", "/docs/examples", Pages.Examples.ExamplesView()
+        | Advanced -> "Advanced", "/docs/advanced", Pages.Advanced.AdvancedView()
     React.router [
         router.hashMode
         router.onUrlChanged (